Update layout for connect screen
