--- lib/screens/login/login_screen.dart
@@ -1,85 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_redux/flutter_redux.dart';
-
-import '../../redux/state/app_state.dart';
-import '../../utils/sentry_colors.dart';
-import 'login_view_model.dart';
-
-class LoginScreen extends StatefulWidget {
-  @override
-  _LoginScreenState createState() => _LoginScreenState();
-}
-
-class _LoginScreenState extends State<LoginScreen> {
-
-  @override
-  Widget build(BuildContext context) {
-    return StoreConnector<AppState, LoginViewModel>(
-      builder: (_, viewModel) => _content(viewModel),
-      converter: (store) => LoginViewModel.fromStore(store),
-    );
-  }
-
-  Widget _content(LoginViewModel viewModel) {
-    return Container(
-        margin: EdgeInsets.all(32.0),
-        child: Center(
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.spaceAround,
-              children: [
-                Column(
-                  children: [
-                    Text(
-                      'Lets get started!',
-                      style: Theme.of(context).textTheme.headline1,
-                      textAlign: TextAlign.center,
-                    ),
-                    SizedBox(height: 12),
-                    Text(
-                      'Login to see your projects.',
-                      style: Theme.of(context).textTheme.subtitle1,
-                      textAlign: TextAlign.center,
-                    ),
-                  ],
-                ),
-                RaisedButton(
-                  child: Text('Connect'),
-                  textColor: Colors.white,
-                  color: SentryColors.rum,
-                  onPressed: () {
-                    setState(() {
-                      // TODO @denis
-                    });
-                  },
-                )
-              ],
-            )
-        )
-    );
-  }
-
-  void _handleLoginFailure(Object error) {
-    showDialog(
-      context: context,
-      barrierDismissible: false,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: Text(
-              'Error',
-              style: Theme.of(context).textTheme.headline1
-          ),
-          content: Text(
-              'Oops, there is a problem. Please try again later.',
-              style: Theme.of(context).textTheme.bodyText1
-          ),
-          actions: [
-            FlatButton(
-                key: Key('ok'),
-                onPressed: () => Navigator.of(context).pop(),
-                child: Text('Ok')),
-          ],
-        );
-      }
-    );
-  }
-}
